vivliostyle/pyproject.toml
[project]
name = "vivliostyle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
